app/pyproject.toml
[project]
name = "personal-assistant-fast-api"
version = "0.1.0"
description = "Fast API for personal assistant agent"
requires-python = "~=3.13.0"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "langgraph-checkpoint-mongodb>=0.1.4",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langgraph-store-mongodb>=0.0.1",
    "nest-asyncio>=1.6.0",
    "psycopg[binary,pool]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "trustcall>=0.0.39",
    "uvicorn>=0.35.0",
]
